requirements.txt
biopython
pandas
python-dotenv
streamlit
plotly
